Fix conflicts related to MyXactAccessedTempRel.
This flag isn't used for anything in GPDB, there is a comment explaining
that in PrepareTransaction. We had #ifdef'd out the variabl, and all
the places that set the variable. To reduce future merge conflicts, I think
it's better to have the variable and keep all the places where it's set
unmodified from the upstream, and only comment out the place where it's
checked in PrepareTransaction. So I did that.