Fix execprocnode probes.
The probe definition in probes.d didn't actually match the PG_TRACE
calls, there was an argument missing. So I guess these never worked quite
right. But the new TRACE_* macros are not as forgiving, and you got a
compiler error. Adjust the definitions in probes.d to match the actual
calls.