Reshuffle a few OIDs to avoid collision.
We had reserved OID 1980 for int8_ops opclass. But that is now used for
numeric_div_trunc function(), which we just merged in. In the upstream, we
have reserved OID 3124 for the opclass, but only since version 9.2. Before
that, we used whatever was free at initdb time. But we have been using OID
3124 for the GPDB-specific pg_proc_callback system table

To resolve this mess, change the OID of pg_proc_callback from 3124 to 7176,
to make 3124 available. And then use 3124 for int8_ops. That leaves 1980
for numeric_div_trunc function(), like in upstream.