Fix the assertion failure from arrays test
This query: "select * from unnest2(array[[1,2,3],[4,5,6]]);" exercises
the code to inline a set-returning function. When inlining a SRF, the
inlined query tree was not "normalized", like the top query tree is.
We had assertions in a few places that checked that a query tree is
fully normalized. Fix by normalizing inlined query trees, as well.