Fix failure in plpgsql test.
ERROR:  unrecognized cmd_type: 7 (pl_funcs.c:618)

In commit cacb2839ac8, we had backported upstream commit 87f2ad1326, which
added the free_stmt() function with a switch-case statement for every
PL/pgSQL statement type. When we backported it, we didn't have
PLPGSQL_STMT_FORC yet, so that was left out of the backport. But now it
was merged in, so we need to add the handler for it in free_stmt(), too.