Fix another failure in plpgsql.
ERROR:  there is no parameter $0
CONTEXT:  SQL statement "update forc_test set i = i * 100, j =  $1  * 2 where current of  $2 "

For some reason, the copy, out, read-functions of CurrentOfExpr were missing
a field. While we're at it, reorder the code to match upstream.

NOTE: That seems like a bug even in GPDB 5. The reason we haven't seen any
failures there is probably that there happens to be no codepaths that would
try to copy a CurrentOfExpr that uses the cursor_param field. Or we're
just missing a test case that would tickle it. Note that if we are to
backport this, we cannot backport the changes to the read/out functions,
because that would require a catalog change if there is a view that
uses it. Although I'm not sure if it's possible to have a view or something
else that contains a CurrentOfExpr in the catalogs.