Allow CURRENT OF on a query with ORDER BY.
Seems to work. I think we rejected it out of sheer conservatism.

Add an ORDER BY to the test query used in the plpgsql test, to make it
pass on GPDB. In GPDB, the rows can come out in any order, so the result
is not repeatable without this. gpdiff handles that for normal selects,
but not for notices like this.

NOTE: We should perhaps add an explicit test case for this, even though
it's now tested by the plpgsql test.