Use LIMIT 0 instead of LIMIT -1 in tests.
Commit bfce56eea4 changed behavior so that a negative limit gives an
error, instead of silently treating it as zero. So don't use -1 in tests,
use 0 instead.

(I'm not sure what the point of these tests was to begin with. In particular,
I'm a bit worried that if the planner was suitably intelligent, it would
notice the LIMIT 0 and generate a plan that doesn't do anything at all. But
I'm not going to worry about that right now, this commit just faithfully
keeps the old behavior.)