Fix ALTER TABLE RENAME TO to work with missing pg_attribute entries again.
If we want to avoid the error about broken catalogs on ALTER TABLE RENAME,
we cannot use the relcache. So the gp_allow_rename_relation_without_lock
GUC isn't really about locking the table, it's about whether to use the
relcache and do the validity checks that the relcache does.

I question how valuable this really is, but I think I had that dicussion
recently on gpdb-dev or somewhere, and some people insisted, so I'm not
going to re-open that can of worms right now. This brings us back to the
current behavior on 'master', but with a smaller diff footprint.