Silence regression failure with ORCA.
ORCA doesn't support this PL/pgSQL cursor to be updatable. I'm not sure why,
but it's not a new issue. (In fact, this used to not work with Postgres
planner either until recently.)

NOTE: Maybe it would be good to have someone investigate this, to see if
it can be fixed in ORCA easily.