Revert "Allow CURRENT OF on a query with ORDER BY.", and fix tests differently
On second thoughts, I think a simpler way to make these tests pass on GPDB
is to add a dummy distribution key column to the test table, so that all
the rows actually reside on a single segment. Supporting ORDER BY is probably
a good idea on its own, but let's not piggyback that on the merge.

This reverts commits 7f8c778f5a13b3505a3cdc0a2f9b9ae24c2f8ace and
fb7b14b75324421a2dc98ffd72ea17cf452d6fb8.