Fix conflicts caused by USE_FLOAT4/FLOAT8_BYVAL changes.
GPDB has rewritten the *GetDatum and DatumGet* macros as static inline
functions. I'm not sure that's a good idea, because of the potential new
compiler warnings that might cause in some places where code is not too
careful about which macro/function to use, and it causes avoidable conflicts
like this. But I kept the GPDB versions for now.

We will not support !USE_FLOAT4/8_BYVAL builds in GPDB.