Remove duplicate entries from pg_proc.
We got the timestamp-variants of generate_series from upstream, as well
as pg_stat_get_activity() function, so we can remove them from pg_proc.sql
now.

pg_stat_get_activity() has some extra OUT arguments in GPDB, so I'm not sure
what the best approach for that would be. I added the extra arguments to
the line in pg_proc.h, but that will conflict with any changes to that line
in the future.