Fixes to tracking error location.
The casts in partition_range_every to A_Const were wrong, the nodes
were actually TypeCast nodes here. That was probably caused by something
we merged in, or this would never have worked. I don't see anything else
in the function that assumes that they're A_Consts, so I guess it's all
right now.

That was still not enough to fix the bogus the error locations in the
'partition' test. We were also missing a COPY_LOCATION_FILED for
TypeCase.