Resolve dblink test output conflicts
The previous DETAIL messages are now dropped and CONTEXT used instead,
which left conflicts since we had ^ markers in the DETAILs. This is no
longer passed which is a shame, but not really something to worry too
much about since dblink ideally will be replaced by postgres_fdw at
some point in the not too distant future.