So called Shared Local Snapshot doesn't contain subxids, reset field.
'subxcnt' was not being initialized. As a result, the subxid array from the
previous snapshot that happened to be in the same struct was reused. This
showed up as occasional failures in the 'appendonly' regression test.
Sometimes, under concurrency, after doing things in a subtransaction and
committing, the next transaction would still consider the subtransaction
as running.