Remove unstable test UPDATE.
This test UPDATE violates two constraints, and it's arbitrary which one it
fails on first. That's problematic, because the error message depends on
which constraint happens to be checked first.

We have tests for testing each failure individually, so testing a case that
fails them both doesn't seem very interesting. We will never reach the
actualy failure of both at the same time, because we bail out on first
error. So to make the output stable, remove the offending UPDATE.

This is https://github.com/greenplum-db/gpdb/pull/3610 on master. I'm pretty
confident that will be merged as is, but if not, then we'll need to sync
iteration_3_redux branch accordingly.