Always copy tuples used in ExecInsert
Scribbling on a shared tuple results in corruption; we need to always be
operating on a local copy. Since commit a77eaa6a95 in the upstream, we
now rely on ExecInsert to make a copy, there is not necessary a junk filter
anymore that makes that for us.

This is a slightly simpler version of Jacob's patch.