WIP: try to recreate the 'materialize slot' logic for heap inserts
This patch tries to recreate a GPDB-specific version of upstream's
ExecMaterializeSlot, which guarantees local storage of the materialized
heaptuple. It does not try to touch memtuples (which may need their own
version of materialization).

We're checking this in to see how the merge CI does. This is not a final
solution, as evidenced by the FIXMEs.