Make qp_dml_oids test case more robust.
I frequently saw failures when running the test on my laptop with ORCA.
The reason is that the OIDs on a user table are not guaranteed to be
unique across segments in GPDB. Depending on concurrent activity and
timing, the OID counters on different segments are not always in sync,
and can produce duplicates when looked at across all nodes. In fact, I
think the only reason this is currently passing on the pipeline so
reliably as it is, is because the test is run in parallel with other
tests that also create objects, which creates enough noise in the OID
allocations.

To fix, modify the test data in the test so that all the initial test rows
reside on the same segment. Within a segment, the OIDs are unique.

NOTE: I submitted this to 'master' as a separate PR:
https://github.com/greenplum-db/gpdb/pull/3735