Fix win32 compilation.
We don't support building GPDB server on Windows, but we do support building
the client programs and libpq. Also, I just noticed that on GPDB,
sizeof(Datum) is always 8, regardless of sizeof(unsigned long). I'm not sure
what would happen if you actually tried to build the server on a 32-bit
system, but let's just keep the behavior the same as before for now.