Another attempt at fixing EXCEPT/INTERSECT.
I tried to fix this earlier by moving the Redistribute node below the
Append, but that hit an assertion, if the subnode of the Append already
contained a Motion. I couldn't immediately figure out how to avoid putting
a Motion on top of a Motion. So instead, keep the Redistribute node above
the Append, but teach it to keep the output of the Motion sorted by the
"flag".