Third attempt at fixing EXCEPT/INTERSECT planning.
The previous attempt added support for a sorted Redistribute motion node,
but that can deadlock. So revert that, and go back to my first approach,
and push the Redistribute Motions to below the Append. But avoid creating
a Motion on top of a Motion, because a) it causes an assertion failure, and
b) it's just silly. I wish we could be smarter and avoid creating useless
Motions in the first place, but that seems like a bigger effort, and isn't
a new issue.