Get src/backend/catalog/heap.c building
A few FIXMEs to check my work on the relpath() calls, and to make sure that our
smgr-related helper functions get updated to call the new RelationXXX() code
that conflicted.

As part of this, fix the signature of RelationTruncate() to match the GPDB
version. storage.c will eventually need to be fixed.