Ok, src/backend/storage/smgr/ compiles now.
All the code to track deletions of relations were moved in upstream from
smgr.c to catalog/storage.c. But we had added tracking for persistent
filespaces, tablespaces and databases there too. Move those as well.

To keep smgr.c clean of GPDB added code, to the extent possible, move the
code to deal with AO tables to new file, smgr_ao.c

md.c can now operate in two modes:

* When working on the main fork, use the Mirrored*() functions, like before
  the merge

* When working on other forks, use File*() functions directly, like in
  upstream.

I also split smgrcreate() function into two: The new smgrmirroredcreate()
must be used to create the main fork. It takes a bunch of extra arguments
compared to upstream, i.e it is equivalent to the old GPDB's smgrcreate()
function. But smgrcreate() now has the same signature as in upstream - it
must be used when creating other forks. This allows us to keep the new FSM
and VM code unchanged, but still use the old code for the main fork.