Revert create_unique_path() function to GPDB version.
This basically means not taking the upstream changes. That decision was
made already in the subselects refactoring, where these commits were
cherry-picked before this merge. It's sad that we can't take the upstream
changes, as that will cause more merge conflicts further down the line,
but I'm not going to try fixing that now.