Fix conflicts on VACUUM and friends.
I cherry-picked commit b503da135a as part of this. We'll get that change
eventually anyway, and adding the 'scanned_all' parameter to vacuum_rel
would've tricky because of the GPDB reorganization of the code. That commit
removes the parameter again, so by cherry-picking it now, we don't need
to reshuffle things to add it, only to remove it later again.