Fix implicit cast warnings
AttrNumber (a signed short) can't handle the OUTER macro (65001). But it
doesn't matter because we don't actually use the initialized value of
the attrno variable here. Get rid of the initialization, and move
declarations into the code block they're used in so it's easier to see
that the resulting logic is correct.