Make sure checksums are computed when calling smgrextend
This is a continued backport of upstream 96ef3b8ff, which was backported before
we had relation forks. This fixes a huge number of checksum warnings around FSM
pages (and probably should fix some visibility map stuff too, but I haven't seen
any messages for that yet).

FIXMEs: we need to backport 20723ce80 so that we can correctly call
MarkBufferDirtyHint() in the FSM code.