Deal with TypeCast expressions appearing in partition boundaries.
This function, as well much of the partition boundary parsing code, ought
to be rewritten. In theory, a typecast of a NULL does not necessarily
yield a NULL value, and on the other hand, a typecast of a non-NULL
constant might yield a NULL value. But I'm not tackling that today.

This function deals with raw expressions before parse analysis, so we
should be using raw_expression_tree_walker here. And
raw_expression_tree_walker() knows how to recurse into TypeCasts, which
solves the immediate problem.