Update expected output for deparsing a function.
I'm not sure why the function call got schema-qualified before. I don't
see any reason for it, and PostgreSQL 8.4 doesn't schema-qualify it, so
let's just update the expected output for it.

NOTE: Much of the 'default_parameters' test is duplicated with
'polymorphism' test. The history is that when the default parameters
feature was cherry-picked to GPDB, the tests that were added to
'polymorphism' were put to 'default_parameters' instead. We should remove
the now-duplicated tests from 'default_parameters' now that we have caught
up.