Don't add a duplicate Material node on top of SubPlans.
This code had been removed in GPDB since time immemorial. It got
accidentally resurreced as part of the merge. Disable it again, but to
prevent that form happening again, comment it out instead of removing it
altogether, and add a comment explaining why it's disabled.

Fixes the 'bfv_joins' failure.