ExecInsert: allow optimizer to issue AFTER ROW triggers
For some reason, my Mac is tripping this assertion, but the CI isn't. I
can only replicate the assertion on CentOS by forcing
optimizer_enable_dml_triggers, so that's what I've done in a
GPDB-specific test.

In any case, since the optimizer might implement foreign key constraints
using a trigger, it seems like we should get rid of this assertion.