Update plan in expected output.
I'm not sure what's changed, but the new plan looks better. There's not
much point in having a HashAggregate just below a Hash Join's Hash node.
Perhaps it would reduce the number of rows that need to be Broadcast, but
the old plan had an extra Redistribute Motion node to facilitate that,
which surely isn't better. Especially when the planner estimated there to
be only one matching row anyway.