Make DynamicBitmapIndexScan a separate node type.
It was inconsistent that DynamicIndexScan and IndexScan were separate node
types, but the dynamic and non-dynamic bitmap index scans were both
represented by BitmapIndexScan. Now they're consistent.

Also, split the IndexScan struct into separate IndexScan and
DynamicIndexScan structs. The logicalIndexInfo field is only used in dynamic
scans, so that goes into the new DynamicIndexScan struct, while the shared
fields are in IndexScan. Similarly for DynamicBitmapIndexScan.

Per Jesse's suggestion.

(This was done in master already, as part of commit 198f701ede.)