Put back ORDER BYs to these test queries.
They are not present in the upstream, but the point of the test is to test
index scans. There's a comment saying:

-- if above selects don't produce the expected tuple order,
-- then you didn't get an indexscan plan, and something is busted.

So it would be weird if the rows don't come out in order. Extra comments
or using EXPLAIN to verify that an index scan was really used would probably
be in order, but for now, let's just avoid changing this as part of the
merge.