Attempt to fix the MU_check_centos failure.
Looks like result[column_key] can be something else than a Python string
here. A more proper fix would probably be to change the mock tests so that
it is a string, but I don't know how to do that. So, just skip the
replacement of :location fields if it's not a string.