Fix misc conflicts.
NOTE: the documentation on window functions in select.sgml is not
up-to-date with what we have in GPDB, since we backported additional
features from later version of PostgreSQL, and have implemented some
extra functionality on top of that. But AFAICS we don't use these SGML
docs for anything, except for the synopsis for psql \h, so just keep it
identical to the current upstream version we're merging, to avoid future
conflicts.