Fix remaining conflicts in src/backend/access/transam.
NOTE: I've resurrected a bunch of code related to archive recovery, that
had been removed in GPDB. Even if we don't support archive recovery, let's
keep the code around to reduce merge conflicts. It's a bit of a mess now,
some code is present, some is missing, and some is present but #ifdef'd
out. But that's the goal.