Resurrect upstream tbm_iterate and introduce tbm_generic_*
If a tbm_* function takes a Node* rather than a specific XxxBitmap*,
rename it to tbm_generic_*. Now the upstream tbm_iterate() (which we
used to call tbm_iterate_hash()) is clearly differentiated.

Note the removal of a comment that made it seem like you could pass NULL
as the second argument to tbm_iterate(). From inspection it seems like
doing that will always cause a segfault; there are no NULL checks before
assigning to that struct's members.

Also remove a dubious call to tbm_generic_upd_instrument(), replacing it
with tbm_upd_instrument().