Fix conflict, by keeping GPDB version.
The retry loop was added in 8.4, but it was removed again in 9.0,
after doing some other changes in src/backend/port/win32/signal.c. We
had the 9.0 version in GPDB, where the retry-loop was already removed. We
do have the signal.c changes already.

All of this is academical on GPDB, of course, as we don't support running
the server on Windows at all.