Move StreamBitmap to the new begin/end_iterate API
Upstream has moved to an API that allows multiple parallel iterations of
a TIDBitmap. This patch attempts to add the same API for StreamBitmaps
(and also adds a thin "generic" wrapper that will choose from the two
implementations depending on what kind of Bitmap is in use). The key
here is to put all mutable iteration state into the StreamBMIterator;
iterating shouldn't touch the StreamBitmap (or the StreamNode) itself.

Notes:
- Not all clients of this code compile yet. A notable case is bitmap.c,
  which creates its own type of StreamNode (sigh...).
- HashStreamOpaque is gone; it wasn't doing enough to justify juggling
  it.
- The begin/end_iterate implementations for StreamBitmap feel too
  copy-pasty; they probably need to be callbacks. This might help with
  the aforementioned bitmap.c.
- I'm not entirely pleased with the StreamNode->opaque field, but
  bitmap.c will probably need *some* sort of state carrier. It seems
  like that should be part of the iterator, though.