bitmap: update to the new StreamNode API
This isn't quite right yet, but it builds. The key problem is that
iterating isn't supposed to modify the StreamNode itself, but we end up
using its IndexScanDesc copy as scratch space during the iteration (even
going so far as to free it). I'm not sure yet what the correct fix is.