Fix expected output for plan change.
This is otherwise the same, but the planner now chooses to do the join of
tt and tq the other way 'round. Not sure why, but the new plan seems at
least as good as the old one, if not better. The new row estimates are more
accurate than those memorized in the old expected output.