Memorize new plan in expected output.
On 'master', you got a normal (= inner) nested loop join, for the subquery
and now you get a semi-join. I think a semi-join is correct, and the new
plan was wrong, although I'm not 100% sure. On 5X_STABLE, you got an
"EXISTS" join, which is equivalent to a semi-join, so even if a semi-join
was not be necessary here, this still wouldn't be a regression compared to
5.X, which is good enough for me.