Implement the NOTICE when REVOKE is a no-op differently.
Instead of throwing a NOTICE on every object/role combination, track
whether anything at all was revoked, and only issue one NOTICE for the
whole command, if nothing was revoked. This reduces the noise if the REVOKE
lists multiple objects and/or roles.

This refactoring makes it easier to carry this diff vs. upstream, as we
merge the column-level permissions feature from upstream.

Discussion: https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/Ttn_UJb4Otg/LS1cFrDiAwAJ

(This is a cherry-pick of PR https://github.com/greenplum-db/gpdb/pull/4050
into iteration_5, to fix a bunch of regression failures)