Ensure we don't call Filerep in the QD
With the change in startup procedure the Checkpointer was calling
FileRep syncs during PMModeMaster where the Filerep shmem segment
isn't set up. Hackily avoid the filerep in QD and UTILITY mode
where we know we don't want to sync file replication. This might
be bogus but it fixes the previously observed segfaults and runs
ICW as fine as can be expected given the state of the iteration.

FIXME added to doublecheck this with CNS.