bitmap.c: allow parallel stream iteration
With the new begin_/end_iterate API, iterators must not touch shared
state. In this case, concurrent iterators were operating on the same
IndexScanDesc. Change stream_begin_iterate() to make a copy of the
IndexScanDesc, and free it during stream_free().

Ownership of the underlying Buffers belongs to the StreamNode. Its free
function, indexstream_free(), is the only one allowed to release them.