Another attempt at normalizing 'date', and related types.
I'm continuing where Jacob left off with previous commit. I don't know how
to run these tests properly, so I'm relying on the pipeline to tell me if
this works or not. It did pass the unit tests on my laptop, that were
failing on ICW jobs in the pipeline with the previous commit. And I test
the modified SQL query manually.