Cherry-pick "Avoid crashing when we have problems unlinking files post-commit."
We had already picked parts of that, more or less by accident, as part of
the merge or the walreplication work. The situation now was that places
had the smgrexists() guards before smgrunlink() calls, but others did not.
mdunlink() was still the old version that required the guards, so those
calls where the smgrexists() guards had already been removed, you now got
warnings about missing FSM and VM files. To fix, cherry-pick the rest of
the upstream commit that removed the smgrexists() calls and made mdunlink()
not complain if extra forks are not found.

This is a cherry-pick of this upstream commit, from PostgreSQL 9.0.7:

commit 61dd2ffaff1be5768151e72aca030d7755255b26
Author: Tom Lane <tgl@sss.pgh.pa.us>
Date:   Tue Dec 20 15:00:47 2011 -0500

    Avoid crashing when we have problems unlinking files post-commit.

    smgrdounlink takes care to not throw an ERROR if it fails to unlink
    something, but that caution was rendered useless by commit
    3396000684b41e7e9467d1abc67152b39e697035, which put an smgrexists call in
    front of it; smgrexists *does* throw error if anything looks funny, such
    as getting a permissions error from trying to open the file.  If that
    happens post-commit, you get a PANIC, and what's worse the same logic
    appears in the WAL replay code, so the database even fails to restart.

    Restore the intended behavior by removing the smgrexists call --- it isn't
    accomplishing anything that we can't do better by adjusting mdunlink's
    ideas of whether it ought to warn about ENOENT or not.

    Per report from Joseph Shraibman of unrecoverable crash after trying to
    drop a table whose FSM fork had somehow gotten chmod'd to 000 permissions.
    Backpatch to 8.4, where the bogus coding was introduced.