Move the logic for how to input for an DML statement into ModifyTable.
This seems to fix some of the common cases, but still lots to be done.

TODOS: Single-Row-Insert optimization is missing. The check for UPDATEing
partition key column is broken.

This now works:
INSERT INTO foo VALUES (123);

But this is still broken:
INSERT INTO foo SELECT 123;

NOTE: The removal of the "pathkeys" arguments from cdbpathtoplan_create_flow
was just coincidental cleanup. I'll submit a separate PR about that..