Fix EXPLAIN of PartitionSelector nodes.
When fixing conflicts earlier, I changed the way the 'parentPlan' is passed
to ExplainNode(). It used to be an additional (compared to upstream)
argument, but now it's passed in the ExplainState struct. I changed that
to minimize the diff vs upstream.

However, I missed this crucial code to actually set the parentPlan field.