Fix serialization of ParamListInfo.
Commit 9bedd128d6 added paramFetch callback function to the ParamListInfo
struct. The old serialization code just copied the in-memory representation
of the struct, and now that included the function pointer. The function
pointer is garbage after serializing and deserializing in a different
process.

To fix, force immediate evaluation of any "lazy" parameters, by calling
the paramFetch hook for every parameter in the QD, before serializing the
params. And don't serialize the function pointers, initialize them to NULL
in the QE.