If we remove a trivial SubQueryScan, keep its Flow.
NOTE: I think we should do this in 5X_STABLE and master too! This showed
up in the regression tests as an assertion failure on this query:

INSERT INTO pg_attribute SELECT distinct on(attrelid, attnum) * FROM gp_dist_random('pg_attribute') WHERE attname='a' AND attrelid='oldcor'::regclass;
FATAL:  Unexpected internal error (nodeMotion.c:251)
DETAIL:  FailedAssertion("!(!"Non-active motion is executed")", File: "nodeMotion.c", Line: 251)

The reason it failed is that the ModifyTable node at the top, on a
master-only table, was attempted to be execued on segments. This used to not
fail, even though we removed the SubqueryScan, because on an INSERT, we set
the Flow of the top node later in the planning, in apply_motion(). But in
principle, I think you could write a query that doesn't even use
INSERT/UPDATE/DELETE, such that removing a trivial SubqueryScan also removes
a Flow that was important.