Fix make_pathkey_from_sortinfo() call
The 'sortref' and 'canonicalize' arguments seem reversed. And why are
we not passing 'tleSortGroupRef' argument, when we have it?

This fixes this error in the 'uao_ddl/temp_on_commit_delete_rows_row' test:
+ERROR:  volatile EquivalenceClass has no sortref (equivclass.c:465)

NOTE: This code is the same in 5X_STABLE. I think it's wrong there as
well, and we ought to fix it, but the same test case seems to work there.