Remove 'exec_eval_cleanup' call that's now wrong.
As noted in the comment, this is indeed not present in PG 9.0. I'm not sure
why it was needed before, but now it was causing these errors in some tests:

ERROR:  cache lookup failed for type 2139062143 (format_type.c:136)

Remove it. If ORCA is still causing trouble, then we'll need to investigate
and fix that some that some other way.