Fix 'foreign_data' test, by cherry-picking a later commit.
The aclexplode function was changed later in the 9.0 release cycle. I'm not
sure why we seem to require that fix, in order to pass the foreign_data
tests, while in the upstream that particular test passed even without it.
But since this fixes it, that's good enough for me.

Upstream commit:

commit 894fc6b763c2053c01a5be5d20a76d1db650547d
Author: Tom Lane <tgl@sss.pgh.pa.us>
Date:   Tue Jan 12 02:39:22 2010 +0000

    Fix aclexplode to not explode on a zero-entry ACL array.
    (An Assert is no substitute for thinking clearly :-()

    Minor style kibitzing too.

    Per report from Robert Treat.