Fix dispatching of queries with record-type parameters.
This fixes the "ERROR:  record type has not been registered" errors we've
been seeing in some tests with PL/pgSQL. With the merge, references like
"r.field" in PL/pgSQL, where 'r' is a record variable, are handled
slightly differently. PL/pgSQL now passes the whole record Datum to the
backend, which executes a FieldSelect on it. But if the query is
dispatched from QD to QE, the QE will not know about the transient record
typmods in the QD process.

This is essentially the same problem we battled with in Motion nodes
in GPDB 5, and added the whole tuple remapper to deal with it. Only this
time, the problem is with record Datums being dispatched from QD to QE,
as Params, rather than with record Datums being transferred across a
Motion.

This is also reproducable on GPDB 5, if you reference the whole record
variable in a query, rather than one field in it. The added regression
test contains an example of that.

NOTE: Because this is a pre-existing, partially, I'll open a separate PR
against master and 5X_STABLE. But pushed now to the iteration branch, to
silence the failures there.