Initial fix for single row insertion optimization.
We just get back of single row insertion optimization. But current
parse tree struture is different and there is a fromlist in it.
Our old logic is generating result node if fromlist is NIL, but planner
generate subqueryscan due to non empty fromlist.

Another fix will come later is we assign gangType as wrong so we execute
const sub select like 'select generate_series(1, 10)' on each segment.
We need to fix it by assign it as GANGTYPE_SINGLETON_READER later.

Heikki: I fixed a few bugs in this Max's commit, and pushed. It doesn't
do UPDATE/DELETE statements yet but it's a step in the right direction.