Remove optimization to "zap" Result nodes above Append nodes.
I'm not convinced that this was ever a meaningful saving. And if it is, we
need to find a new way of reaping that benefit. As the code stands, it was
causing incorrect results in queries in 'partition' test, where the
optimization kicked in.

NOTE: This probably warrants a separate discussion on gpdb-dev, and we
can commit this as a separate PR.