Revert the earlier "zapping" related changes, and do PR #4448 instead
I proposed removing the "zappign" code in:

https://github.com/greenplum-db/gpdb/pull/4448

This commit changes the code in iteration_6 to match that PR. While we may
still have some chanegs to that PR before it's pushed, let's try to keep
'iteration_6' in sync with it, to avoid getting confused.

Compared to what we have on iteration_6 at the moment, that PR adds code
to plan_pushdown_tlist() to avoid inserting unnecessary Result nodes.